refactor: move main
