+++ __tests__/report/golangcilint.test.ts
@@ -1,4 +1,4 @@
-import { Result } from '../../src/type'
+import { Result } from '../../src/common/type'
 import { GolangCILintSummaryReportImpl } from '../../src/report/golangcilint'
 
 describe('GolangCILintSummaryImpl', () => {
