+++ src/input.ts
@@ -3,8 +3,8 @@ import * as github from '@actions/github'
 import YAML from 'yaml'
 
 import { ConfigSchema } from './config.generated'
-import { XmlFileGroup } from '../table/factory'
-import { GitHubContext, ReporterType } from '../common/type'
+import { XmlFileGroup } from './table/factory'
+import { GitHubContext, ReporterType } from './common/type'
 
 export function getGitHubToken(): string {
   return core.getInput('github-token', { required: true })
